feat: use dotnet-deps-parser
For .NET Framework projects, if the targetFramework cannot be determined, try to use the dotnet-deps-parser as a last resort.
This fixes a problem where only a packages.config file is present in the directory.
Currently the implementation looks only for .csproj files, so this attempts to align the plugin with how docker-deps works (with all .NET files).